Add ShoppingCart Table & CartItem Table
When user is created automatically generates a new shopping cart to go
along with them. That new shopping cart add the users_entity

+++ caseStudySpringBootv2/src/main/java/com/controller/ProductEntityController.java
@@ -32,7 +32,7 @@ public class ProductEntityController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity updateProduct(@PathVariable Long id, @RequestBody ProductEntity product) {
+    public ResponseEntity updateProduct(@PathVariable Integer id, @RequestBody ProductEntity product) {
         ProductEntity currentProduct = productRepository.findById(id).orElseThrow(RuntimeException::new);
         currentProduct.setName(product.getName());
         currentProduct.setDescription(product.getDescription());
@@ -42,7 +42,7 @@ public class ProductEntityController {
     }
 
     @DeleteMapping("/{id")
-    public ResponseEntity deleteProduct(@PathVariable long id) {
+    public ResponseEntity deleteProduct(@PathVariable Integer id) {
         productRepository.deleteById(id);
         return ResponseEntity.ok().build();
     }
